Begin creation of NEI class
This commit includes the groundwork for a non-equilibrium ionization
(NEI) modeling class.  There is still much work to be done.  Right
now, this branch depends on the ionization-state branch of
namurphy/PlasmaPy which has the IonizationState and IonizationStates
classes which have not yet been merged (see PlasmaPy/PlasmaPy#404).
Much of what I put in so far has been to set up the setters and
getters for necessary attributes for the class.  The time_advance and
simulation methods still need to be added.