Temporarily put IonizationState(s) in this package
These classes are currently in the ionization-state of
namurphy/PlasmaPy.  However, since they are not in an official release
of PlasmaPy yet, it will cause fewer problems to just put them here.
This way, we should be able to just do a `pip install plasmapy` and
get version 0.1.1, which will have all of the needed functionality.