Update docstrings and type hints in NEI class
Right now for type hint annotations, I am using astropy.units.Quantity
for Quantity instances, though I had just before this used the units
that a quantity had to be in.  However, the difficulty with this is
that units are not types, whereas Quantity is a type.  I was therefore
not able to do things like Dict[str, astropy.units.s].  I wonder if it
would be possible to do something like Quantity[s] which could return
Quantity itself.  Overall, I'm not sure how Quantity instances with
units fit into type hint annotations.